+++ src/server/devices/zigbee/zigbeeIlluLedRGBCCT.ts
@@ -1,17 +1,13 @@
 import { DeviceType } from '../deviceType';
 import { LedSettings, LogLevel, TimeOfDay } from '../../../models';
-import { ZigbeeDevice } from './BaseDevices';
 import { IoBrokerDeviceInfo } from '../IoBrokerDeviceInfo';
+import { ZigbeeIlluDimmer } from './zigbeeIlluDimmer';
 
-export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
+export class ZigbeeIlluLedRGBCCT extends ZigbeeIlluDimmer {
   public static DEFAULT_COLOR_WARM: string = '#f2b200';
-  public on: boolean = false;
-  public brightness: number = 0;
   public color: string = '#fcba32';
   public colortemp: number = 500;
-  public settings: LedSettings = new LedSettings();
-  private stateID: string = 'state';
-  private brightnessID: string = 'brightness';
+  public override settings: LedSettings = new LedSettings();
   private colorID: string = '';
   private colorTempID: string = '';
 
@@ -19,8 +15,6 @@ export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
 
   public constructor(pInfo: IoBrokerDeviceInfo) {
     super(pInfo, DeviceType.ZigbeeIlluLedRGBCCT);
-    this.stateID = `${this.info.fullID}.state`;
-    this.brightnessID = `${this.info.fullID}.brightness`;
     this.colorID = `${this.info.fullID}.color`;
     this.colorTempID = `${this.info.fullID}.colortemp`;
     // this.effectID = `${this.info.fullID}.effect`;
@@ -28,16 +22,8 @@ export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
 
   public update(idSplit: string[], state: ioBroker.State, initial: boolean = false): void {
     this.log(LogLevel.DeepTrace, `LED Update: ID: ${idSplit.join('.')} JSON: ${JSON.stringify(state)}`);
-    super.update(idSplit, state, initial, true);
+    super.update(idSplit, state, initial);
     switch (idSplit[3]) {
-      case 'state':
-        this.log(LogLevel.Trace, `LED Update für ${this.info.customName} auf ${state.val}`);
-        this.on = state.val as boolean;
-        break;
-      case 'brightness':
-        this.log(LogLevel.Trace, `LED Helligkeit Update für ${this.info.customName} auf ${state.val}`);
-        this.brightness = state.val as number;
-        break;
       case 'color':
         this.log(LogLevel.Trace, `LED Color Update für ${this.info.customName} auf ${state.val}`);
         this.color = state.val as string;
@@ -49,32 +35,72 @@ export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
     }
   }
 
-  public setTimeBased(time: TimeOfDay): void {
+  public override setTimeBased(time: TimeOfDay, timeout: number = -1, force: boolean = false): void {
     switch (time) {
       case TimeOfDay.Night:
         if (this.settings.nightOn) {
-          this.setLight(true, this.settings.nightBrightness, this.settings.nightColor, this.settings.nightColorTemp);
+          this.setLight(
+            true,
+            timeout,
+            force,
+            this.settings.nightBrightness,
+            undefined,
+            this.settings.nightColor,
+            this.settings.nightColorTemp,
+          );
         }
         break;
       case TimeOfDay.AfterSunset:
         if (this.settings.duskOn) {
-          this.setLight(true, this.settings.duskBrightness, this.settings.duskColor, this.settings.duskColorTemp);
+          this.setLight(
+            true,
+            timeout,
+            force,
+            this.settings.duskBrightness,
+            undefined,
+            this.settings.duskColor,
+            this.settings.duskColorTemp,
+          );
         }
         break;
       case TimeOfDay.BeforeSunrise:
         if (this.settings.dawnOn) {
-          this.setLight(true, this.settings.dawnBrightness, this.settings.dawnColor, this.settings.dawnColorTemp);
+          this.setLight(
+            true,
+            timeout,
+            force,
+            this.settings.dawnBrightness,
+            undefined,
+            this.settings.dawnColor,
+            this.settings.dawnColorTemp,
+          );
         }
         break;
       case TimeOfDay.Daylight:
         if (this.settings.dayOn) {
-          this.setLight(true, this.settings.dayBrightness, this.settings.dayColor, this.settings.dayColorTemp);
+          this.setLight(
+            true,
+            timeout,
+            force,
+            this.settings.dayBrightness,
+            undefined,
+            this.settings.dayColor,
+            this.settings.dayColorTemp,
+          );
         }
         break;
     }
   }
 
-  public setLight(pValue: boolean, brightness: number = -1, color: string = '', colortemp: number = -1): void {
+  public override setLight(
+    pValue: boolean,
+    timeout?: number,
+    force?: boolean,
+    brightness: number = -1,
+    transitionTime?: number,
+    color: string = '',
+    colorTemp: number = -1,
+  ): void {
     if (this.stateID === '') {
       this.log(LogLevel.Error, `Keine State ID bekannt.`);
       return;
@@ -90,8 +116,9 @@ export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
     }
     this.log(
       LogLevel.Debug,
-      `LED Schalten An: ${pValue}\tHelligkeit: ${brightness}%\tFarbe: "${color}"\tColorTemperatur: ${colortemp}`,
+      `LED Schalten An: ${pValue}\tHelligkeit: ${brightness}%\tFarbe: "${color}"\tColorTemperatur: ${colorTemp}`,
     );
+    super.setLight(pValue, timeout, force, brightness, transitionTime);
 
     if (color !== '') {
       this.ioConn.setState(this.colorID, color, (err) => {
@@ -101,8 +128,8 @@ export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
       });
     }
 
-    if (colortemp > -1) {
-      this.ioConn.setState(this.colorTempID, colortemp, (err) => {
+    if (colorTemp > -1) {
+      this.ioConn.setState(this.colorTempID, colorTemp, (err) => {
         if (err) {
           this.log(LogLevel.Error, `LED Farbwärme schalten ergab Fehler: ${err}`);
         }
@@ -114,19 +141,5 @@ export class ZigbeeIlluLedRGBCCT extends ZigbeeDevice {
         this.log(LogLevel.Error, `LED schalten ergab Fehler: ${err}`);
       }
     });
-
-    if (brightness > -1) {
-      this.ioConn.setState(this.brightnessID, brightness, (err) => {
-        if (err) {
-          this.log(LogLevel.Error, `LED Helligkeit schalten ergab Fehler: ${err}`);
-        }
-      });
-    }
-  }
-
-  public toggleLight(): boolean {
-    const newVal = !this.on;
-    this.setLight(newVal);
-    return newVal;
   }
 }