+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Zigbee IlluLedRGBCCT as Dimmable Lamp
+
 ## 1.1.19 (2022-09-17)
 
 * (theimo1221) Fix Room Temperature naming